cygwin build for examples, fixed warnings for quantumkeeper

+++ include/tlm_h/tlm_quantum/tlm_quantumkeeper.h
@@ -157,7 +157,7 @@ namespace tlm {
         if (mGlobalQuantum != sc_core::SC_ZERO_TIME) {
           const sc_core::sc_time currentTime = sc_core::sc_time_stamp();
           sc_dt::int64 tmp = static_cast<sc_dt::int64>(currentTime / mGlobalQuantum);
-          const sc_core::sc_time remainder = (tmp + 1) * mGlobalQuantum - currentTime;
+          const sc_core::sc_time remainder = (double)(tmp + 1) * mGlobalQuantum - currentTime;
           return remainder;
   
         } else {
